computeStandardizedDifference {FeatureExtraction}	R Documentation

Compute standardized difference of mean for all covariates.

Description

Computes the standardized difference for all covariates between two cohorts. The standardized difference is defined as the difference between the mean divided by the overall standard deviation.

Usage

computeStandardizedDifference(
  covariateData1,
  covariateData2,
  cohortId1 = NULL,
  cohortId2 = NULL
)

Arguments

covariateData1
The covariate data of the first cohort. Needs to be in aggregated format.

covariateData2
The covariate data of the second cohort. Needs to be in aggregated format.

cohortId1
If provided, covariateData1 will be restricted to this cohort. If not provided, covariateData1 is assumed to contain data on only 1 cohort.

cohortId2
If provided, covariateData2 will be restricted to this cohort. If not provided, covariateData2 is assumed to contain data on only 1 cohort.

Value

A data frame with means and standard deviations per cohort as well as the standardized difference of mean.

[Package FeatureExtraction version 3.1.1 Index]