createTable1CovariateSettings {FeatureExtraction}	R Documentation

Create covariate settings for a table 1

Description

Creates a covariate settings object for generating only those covariates that will be included in a table 1. This function works by filtering the covariateSettings object for the covariates in the specifications object.

Usage

createTable1CovariateSettings(
  specifications = getDefaultTable1Specifications(),
  covariateSettings = createDefaultCovariateSettings(),
  includedCovariateConceptIds = c(),
  addDescendantsToInclude = FALSE,
  excludedCovariateConceptIds = c(),
  addDescendantsToExclude = FALSE,
  includedCovariateIds = c()
)

Arguments

specifications
A specifications object for generating a table using the createTable1 function.

covariateSettings
The covariate settings object to use as the basis for the filtered covariate settings.

includedCovariateConceptIds
A list of concept IDs that should be used to construct covariates.

addDescendantsToInclude
Should descendant concept IDs be added to the list of concepts to include?

excludedCovariateConceptIds
A list of concept IDs that should NOT be used to construct covariates.

addDescendantsToExclude
Should descendant concept IDs be added to the list of concepts to exclude?

includedCovariateIds
A list of covariate IDs that should be restricted to.

Value

A covariate settings object, for example to be used when calling the getDbCovariateData function.

[Package FeatureExtraction version 3.1.1 Index]