loadCovariateData {FeatureExtraction}	R Documentation

Load the covariate data from a folder

Description

loadCovariateData loads an object of type covariateData from a folder in the file system.

Usage

loadCovariateData(file, readOnly)

Arguments

file
The name of the folder containing the data.

readOnly
DEPRECATED: If true, the data is opened read only.

Details

The data will be written to a set of files in the folder specified by the user.

Value

An object of class CovariateData.

Examples

# todo

[Package FeatureExtraction version 3.1.1 Index]