saveCovariateData {FeatureExtraction}	R Documentation

Save the covariate data to folder

Description

saveCovariateData saves an object of type covariateData to folder.

Usage

saveCovariateData(covariateData, file)

Arguments

covariateData
An object of type covariateData as generated using getDbCovariateData.

file
The name of the folder where the data will be written. The folder should not yet exist.

Details

The data will be written to a set of files in the folder specified by the user.

Examples

# todo

[Package FeatureExtraction version 3.1.1 Index]