CovariateData-class {FeatureExtraction}	R Documentation

Covariate Data

Description

CovariateData is an S4 class that inherits from Andromeda. It contains information on covariates, which can be either captured on a per-person basis, or aggregated across the cohort(s).

By default covariates refer to a specific time period, with for example different covariate IDs for whether a diagnosis code was observed in the year before and month before index date. However, a CovariateData can also be temporal, meaning that next to a covariate ID there is also a time ID, which identifies the (user specified) time window the covariate was captured.

A CovariateData object is typically created using getDbCovariateData, can only be saved using saveCovariateData, and loaded using loadCovariateData.

Usage

## S4 method for signature 'CovariateData'
show(object)

## S4 method for signature 'CovariateData'
summary(object)

Arguments

object
An object of class 'CovariateData'.

See Also

isCovariateData, isAggregatedCovariateData, isTemporalCovariateData

[Package FeatureExtraction version 3.2.0 Index]