createTable1 {FeatureExtraction}	R Documentation

Create a table 1

Description

Creates a formatted table of cohort characteristics, to be included in publications or reports. Allows for creating a table describing a single cohort, or a table comparing two cohorts.

Usage

createTable1(
  covariateData1,
  covariateData2 = NULL,
  cohortId1 = NULL,
  cohortId2 = NULL,
  specifications = getDefaultTable1Specifications(),
  output = "two columns",
  showCounts = FALSE,
  showPercent = TRUE,
  percentDigits = 1,
  valueDigits = 1,
  stdDiffDigits = 2
)

Arguments

covariateData1
The covariate data of the cohort to be included in the table.

covariateData2
The covariate data of the cohort to also be included, when comparing two cohorts.

cohortId1
If provided, covariateData1 will be restricted to this cohort. If not provided, covariateData1 is assumed to contain data on only 1 cohort.

cohortId2
If provided, covariateData2 will be restricted to this cohort. If not provided, covariateData2 is assumed to contain data on only 1 cohort.

specifications
Specifications of which covariates to display, and how.

output
The output format for the table. Options are output = "two columns", output = "one column", or output = "list".

showCounts
Show the number of cohort entries having the binary covariate?

showPercent
Show the percentage of cohort entries having the binary covariate?

percentDigits
Number of digits to be used for percentages.

valueDigits
Number of digits to be used for the values of continuous variables.

stdDiffDigits
Number of digits to be used for the standardized differences.

Value

A data frame, or, when output = "list" a list of two data frames.

[Package FeatureExtraction version 3.1.1 Index]