Create covariate settings

Create covariate settings

createTemporalCovariateSettings(
  useDemographicsGender = FALSE,
  useDemographicsAge = FALSE,
  useDemographicsAgeGroup = FALSE,
  useDemographicsRace = FALSE,
  useDemographicsEthnicity = FALSE,
  useDemographicsIndexYear = FALSE,
  useDemographicsIndexMonth = FALSE,
  useDemographicsPriorObservationTime = FALSE,
  useDemographicsPostObservationTime = FALSE,
  useDemographicsTimeInCohort = FALSE,
  useDemographicsIndexYearMonth = FALSE,
  useConditionOccurrence = FALSE,
  useConditionOccurrencePrimaryInpatient = FALSE,
  useConditionEraStart = FALSE,
  useConditionEraOverlap = FALSE,
  useConditionEraGroupStart = FALSE,
  useConditionEraGroupOverlap = FALSE,
  useDrugExposure = FALSE,
  useDrugEraStart = FALSE,
  useDrugEraOverlap = FALSE,
  useDrugEraGroupStart = FALSE,
  useDrugEraGroupOverlap = FALSE,
  useProcedureOccurrence = FALSE,
  useDeviceExposure = FALSE,
  useMeasurement = FALSE,
  useMeasurementValue = FALSE,
  useMeasurementRangeGroup = FALSE,
  useObservation = FALSE,
  useCharlsonIndex = FALSE,
  useDcsi = FALSE,
  useChads2 = FALSE,
  useChads2Vasc = FALSE,
  useHfrs = FALSE,
  useDistinctConditionCount = FALSE,
  useDistinctIngredientCount = FALSE,
  useDistinctProcedureCount = FALSE,
  useDistinctMeasurementCount = FALSE,
  useDistinctObservationCount = FALSE,
  useVisitCount = FALSE,
  useVisitConceptCount = FALSE,
  temporalStartDays = -365:-1,
  temporalEndDays = -365:-1,
  includedCovariateConceptIds = c(),
  addDescendantsToInclude = FALSE,
  excludedCovariateConceptIds = c(),
  addDescendantsToExclude = FALSE,
  includedCovariateIds = c()
)

Arguments

useDemographicsGender
Gender of the subject. (analysis ID 1)

useDemographicsAge
Age of the subject on the index date (in years). (analysis ID 2)

useDemographicsAgeGroup
Age of the subject on the index date (in 5 year age groups) (analysis ID 3)

useDemographicsRace
Race of the subject. (analysis ID 4)

useDemographicsEthnicity
Ethnicity of the subject. (analysis ID 5)

useDemographicsIndexYear
Year of the index date. (analysis ID 6)

useDemographicsIndexMonth
Month of the index date. (analysis ID 7)

useDemographicsPriorObservationTime
Number of days of observation time preceding the index date. (analysis ID 8)

useDemographicsPostObservationTime
Number of days of observation time preceding the index date. (analysis ID 9)

useDemographicsTimeInCohort
Number of days of observation time preceding the index date. (analysis ID 10)

useDemographicsIndexYearMonth
Calendar month of the index date. (analysis ID 11)

useConditionOccurrence
One covariate per condition in the condition_occurrence table starting in the time window. (analysis ID 101)

useConditionOccurrencePrimaryInpatient
One covariate per condition observed as a primary diagnosis in an inpatient setting in the condition_occurrence table starting in the time window. (analysis ID 102)

useConditionEraStart
One covariate per condition in the condition_era table starting in the time window. (analysis ID 201)

useConditionEraOverlap
One covariate per condition in the condition_era table overlapping with any part of the time window. (analysis ID 202)

useConditionEraGroupStart
One covariate per condition era rolled up to SNOMED groups in the condition_era table starting in the time window. (analysis ID 203)

useConditionEraGroupOverlap
One covariate per condition era rolled up to SNOMED groups in the condition_era table overlapping with any part of the time window. (analysis ID 203)

useDrugExposure
One covariate per drug in the drug_exposure table starting in the time window. (analysis ID 301)

useDrugEraStart
One covariate per drug in the drug_era table starting in the time window. (analysis ID 401)

useDrugEraOverlap
One covariate per drug in the drug_era table overlapping with any part of the time window. (analysis ID 402)

useDrugEraGroupStart
One covariate per drug rolled up to ATC groups in the drug_era table starting in thetime window. (analysis ID 403)

useDrugEraGroupOverlap
One covariate per drug rolled up to ATC groups in the drug_era table overlapping with any part of thetime window. (analysis ID 403)

useProcedureOccurrence
One covariate per procedure in the procedure_occurrence table in the time window. (analysis ID 501)

useDeviceExposure
One covariate per device in the device exposure table starting in the timewindow. (analysis ID 601)

useMeasurement
One covariate per measurement in the measurement table in the time window. (analysis ID 701)

useMeasurementValue
One covariate containing the value per measurement-unit combination in the time window. If multiple values are found, the last is taken. (analysis ID 702)

useMeasurementRangeGroup
Covariates indicating whether measurements are below, within, or above normal range within the time period. (analysis ID 703)

useObservation
One covariate per observation in the observation table in the time window. (analysis ID 801)

useCharlsonIndex
The Charlson comorbidity index (Romano adaptation) using all conditions prior to the window end. (analysis ID 901)

useDcsi
The Diabetes Comorbidity Severity Index (DCSI) using all conditions prior to the window end. (analysis ID 902)

useChads2
The CHADS2 score using all conditions prior to the window end. (analysis ID 903)

useChads2Vasc
The CHADS2VASc score using all conditions prior to the window end. (analysis ID 904)

useHfrs
The Hospital Frailty Risk Score score using all conditions prior to the window end. (analysis ID 926)

useDistinctConditionCount
The number of distinct condition concepts observed in the time window. (analysis ID 905)

useDistinctIngredientCount
The number of distinct ingredients observed in the time window. (analysis ID 906)

useDistinctProcedureCount
The number of distinct procedures observed in the time window. (analysis ID 907)

useDistinctMeasurementCount
The number of distinct measurements observed in the time window. (analysis ID 908)

useDistinctObservationCount
The number of distinct observations in the time window. (analysis ID 909)

useVisitCount
The number of visits observed in the time window. (analysis ID 910)

useVisitConceptCount
The number of visits observed in the time window, stratified by visit concept ID. (analysis ID 911)

temporalStartDays
A list of integers representing the start of a time period, relative to the index date. 0 indicates the index date, -1 indicates the day before the index date, etc. The start day is included in the time period.

temporalEndDays
A list of integers representing the end of a time period, relative to the index date. 0 indicates the index date, -1 indicates the day before the index date, etc. The end day is included in the time period.

includedCovariateConceptIds
A list of concept IDs that should be used to construct covariates.

addDescendantsToInclude
Should descendant concept IDs be added to the list of concepts to include?

excludedCovariateConceptIds
A list of concept IDs that should NOT be used to construct covariates.

addDescendantsToExclude
Should descendant concept IDs be added to the list of concepts to exclude?

includedCovariateIds
A list of covariate IDs that should be restricted to.

Value

An object of type covariateSettings, to be used in other functions.

Details

creates an object specifying how covariates should be contructed from data in the CDM model.

Examples

settings <- createTemporalCovariateSettings(useDemographicsGender = TRUE,
                                            useDemographicsAge = FALSE,
                                            useDemographicsAgeGroup = TRUE,
                                            useDemographicsRace = TRUE,
                                            useDemographicsEthnicity = TRUE,
                                            useDemographicsIndexYear = TRUE,
                                            useDemographicsIndexMonth = TRUE,
                                            useDemographicsPriorObservationTime = FALSE,
                                            useDemographicsPostObservationTime = FALSE,
                                            useDemographicsTimeInCohort = FALSE,
                                            useDemographicsIndexYearMonth = FALSE,
                                            useConditionOccurrence = FALSE,
                                            useConditionOccurrencePrimaryInpatient = FALSE,
                                            useConditionEraStart = FALSE,
                                            useConditionEraOverlap = FALSE,
                                            useConditionEraGroupStart = FALSE,
                                            useConditionEraGroupOverlap = TRUE,
                                            useDrugExposure = FALSE,
                                            useDrugEraStart = FALSE,
                                            useDrugEraOverlap = FALSE,
                                            useDrugEraGroupStart = FALSE,
                                            useDrugEraGroupOverlap = TRUE,
                                            useProcedureOccurrence = TRUE,
                                            useDeviceExposure = TRUE,
                                            useMeasurement = TRUE,
                                            useMeasurementValue = FALSE,
                                            useMeasurementRangeGroup = TRUE,
                                            useObservation = TRUE,
                                            useCharlsonIndex = TRUE,
                                            useDcsi = TRUE,
                                            useChads2 = TRUE,
                                            useChads2Vasc = TRUE,
                                            useHfrs = FALSE,
                                            useDistinctConditionCount = FALSE,
                                            useDistinctIngredientCount = FALSE,
                                            useDistinctProcedureCount = FALSE,
                                            useDistinctMeasurementCount = FALSE,
                                            useDistinctObservationCount = FALSE,
                                            useVisitCount = FALSE,
                                            useVisitConceptCount = FALSE,
                                            temporalStartDays = -365:-1,
                                            temporalEndDays = -365:-1,
                                            includedCovariateConceptIds = c(),
                                            addDescendantsToInclude = FALSE,
                                            excludedCovariateConceptIds = c(),
                                            addDescendantsToExclude = FALSE,
                                            includedCovariateIds = c())