tidyCovariateData {FeatureExtraction}	R Documentation

Tidy covariate data

Description

Tidy covariate data

Usage

tidyCovariateData(
  covariateData,
  minFraction = 0.001,
  normalize = TRUE,
  removeRedundancy = TRUE
)

Arguments

covariateData
An object as generated using the getDbCovariateData function.

minFraction
Minimum fraction of the population that should have a non-zero value for a covariate for that covariate to be kept. Set to 0 to don't filter on frequency.

normalize
Normalize the covariates? (dividing by the max).

removeRedundancy
Should redundant covariates be removed?

Details

Normalize covariate values by dividing by the max and/or remove redundant covariates and/or remove infrequent covariates. For temporal covariates, redundancy is evaluated per time ID.

[Package FeatureExtraction version 3.1.1 Index]